pyproject.toml
[project]
name = "quipu-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "groq>=0.37.1",
    "httpx>=0.28.1",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.8",
    "matplotlib>=3.10.8",
    "orjson>=3.11.7",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "structlog>=25.5.0",
    "supabase>=2.28.0",
    "tenacity>=9.1.4",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
